+++ src/telemetry/tls_config_reload.rs
@@ -7,8 +7,9 @@ use std::{
 
 use telemetry::{
     metrics::{
-        prom::{FmtLabels, FmtMetrics},
         Counter,
+        FmtLabels,
+        FmtMetrics,
         Gauge,
         Scopes,
     },
@@ -40,7 +41,7 @@ pub fn new() -> (Sensor, Report) {
 pub struct Sensor(Arc<Mutex<Inner>>);
 
 /// Formats metrics for Prometheus for a corresonding `Sensor`.
-#[derive(Debug, Default)]
+#[derive(Clone, Debug, Default)]
 pub struct Report(Weak<Mutex<Inner>>);
 
 #[derive(Debug, Default)]
